refactor(mesh): give startup readiness its own module
The review fix pushed `commands/mesh_llm.rs` from 994 to 1023 lines, over the
1000-line ceiling the desktop ratchet enforces — a file already over the limit
may not grow, and this one was three lines under it.

Readiness is a separable responsibility: probe the chat path agents actually
use, and classify a timeout into copy that names the stage. It moves whole,
with its tests, into `commands/mesh_readiness.rs`. No behaviour change.

  mesh_llm.rs        1023 -> 858
  mesh_readiness.rs         251 (with its 7 tests)

The tests move with the code they cover rather than staying behind in
`mesh_llm_tests.rs`, so the module owns its own contract.
